POK-65 refactor navigation to use single activity
