+++ PokeAppAndroid/Resources/layout/fragment_login.xml
@@ -1,0 +1,152 @@
+﻿<?xml version="1.0" encoding="UTF-8" ?>
+<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"    
+    android:background="@drawable/gradient_background">
+
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:id="@+id/constraintLayoutId"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="vertical"
+        android:layout_gravity="center_vertical"
+        android:layout_marginTop="35dp">
+
+		<LinearLayout
+			android:layout_width="match_parent"
+			android:layout_height="wrap_content"
+			android:id="@+id/viewLayout"
+			app:layout_constraintTop_toBottomOf="@id/pikachuImageView"
+			android:orientation="vertical"
+            android:gravity="center"
+            app:layout_constraintWidth_percent="0.7"
+            app:layout_constraintHeight_percent="0.7"
+            >
+
+	        <ImageView
+		        android:id="@+id/titleImageView"
+                style="@style/DefaultMargin"
+                app:layout_constraintWidth_percent="0.5"
+                app:layout_constraintHeight_percent="0.5"            
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintEnd_toEndOf="parent"
+		        android:layout_width="wrap_content"
+		        android:layout_height="wrap_content"
+                android:layout_marginTop="10dp"
+		        android:src="@drawable/title"
+		        app:layout_constraintTop_toTopOf="parent"
+		        app:layout_constraintBottom_toTopOf="@id/pikachuImageView"
+		        app:layout_constraintVertical_chainStyle="packed"
+		        android:contentDescription="@string/pikachuImage" 
+            />
+
+	        <ImageView
+		        android:id="@+id/pikachuImageView"
+                style="@style/DefaultImageMarginTop"
+                app:layout_constraintWidth_percent="0.5"
+                app:layout_constraintHeight_percent="0.5"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintEnd_toEndOf="parent"
+		        android:layout_width="wrap_content"
+		        android:layout_height="wrap_content"
+                android:layout_marginBottom="10dp"
+		        android:src="@drawable/pikachu"
+		        app:layout_constraintTop_toBottomOf="@id/titleImageView"
+		        app:layout_constraintBottom_toTopOf="@id/viewLayout"
+		        app:layout_constraintVertical_chainStyle="packed"
+		        android:contentDescription="@string/pikachuTitle" 
+                />
+
+            <LinearLayout
+                style="@style/DefaultImageMarginTop"
+                android:layout_width="match_parent"
+			    android:layout_height="100px"
+			    android:id="@+id/viewLayout"
+			    android:orientation="horizontal"
+                android:gravity="center">
+
+                 <TextView
+                    android:id="@+id/usernameTextView"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:text="@string/usernameText"
+                    android:textSize="18sp"
+                    app:layout_constraintTop_toBottomOf="@id/viewLayout"
+                    app:layout_constraintStart_toStartOf="parent"
+                    app:layout_constraintEnd_toStartOf="@id/usernameEditText"
+                    app:layout_constraintHorizontal_bias="0.0"
+                    app:layout_constraintVertical_bias="0.0"
+                 />
+
+                <EditText
+                    android:id="@+id/usernameEditText"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:inputType="text"
+                    android:padding="10dp"
+                    android:hint="@string/usernamePlaceholder"
+                    app:layout_constraintTop_toBottomOf="@id/viewLayout"
+                    app:layout_constraintStart_toEndOf="@id/usernameTextView"
+                    app:layout_constraintEnd_toEndOf="parent"
+                    app:layout_constraintHorizontal_bias="0.0"
+                    app:layout_constraintVertical_bias="0.0"
+                />
+            </LinearLayout>
+
+            <LinearLayout                
+                style="@style/DefaultImageMarginTop"
+                android:layout_width="match_parent"
+			    android:layout_height="100px"
+			    android:id="@+id/viewLayout2"
+			    android:orientation="horizontal" 
+                app:layout_constraintTop_toBottomOf="@id/viewLayout"
+                android:gravity="center">>
+
+                <TextView
+                    android:id="@+id/passwordTextView"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:text="@string/passwordText"
+                    android:textSize="18sp"
+                    app:layout_constraintTop_toBottomOf="@id/usernameEditText"
+                    app:layout_constraintStart_toStartOf="parent"
+                    app:layout_constraintEnd_toStartOf="@id/passwordEditText"
+                    app:layout_constraintHorizontal_bias="0.0"
+                    app:layout_constraintVertical_bias="0.0"
+                />
+				<EditText
+					android:id="@+id/passwordEditText"
+					android:layout_width="wrap_content"
+					android:layout_height="wrap_content"
+					android:inputType="textWebPassword"
+					android:hint="@string/passwordPlaceholder"
+					android:padding="10dp"
+					app:layout_constraintTop_toBottomOf="@id/usernameEditText"
+					app:layout_constraintStart_toEndOf="@id/passwordTextView"
+					app:layout_constraintEnd_toEndOf="parent"
+					app:layout_constraintHorizontal_bias="0.0"
+					app:layout_constraintVertical_bias="0.0"
+                />
+            </LinearLayout>
+ 
+		</LinearLayout>
+
+        <Button
+            android:id="@+id/loginButton"
+            android:layout_marginTop="10dp"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="@string/login_title"
+            android:textColor="#747474"
+            style="@style/MyButtonStyle"
+            app:layout_constraintTop_toBottomOf="@id/viewLayout"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintVertical_bias="0.0"
+            app:layout_constraintHorizontal_bias="0.5"
+            app:layout_constraintWidth_percent="0.7"
+        />
+    </androidx.constraintlayout.widget.ConstraintLayout>
+</ScrollView>
+
