--- PokeAppAndroid/View/LoginActivity.cs
@@ -1,27 +1,0 @@
-﻿
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-using Android.App;
-using Android.Content;
-using Android.OS;
-using Android.Runtime;
-using Android.Views;
-using Android.Widget;
-
-namespace PokeAppAndroid.View
-{
-	[Activity (Label = "LoginActivity")]			
-	public class LoginActivity : Activity
-	{
-		protected override void OnCreate (Bundle savedInstanceState)
-		{
-			base.OnCreate (savedInstanceState);
-
-			// Create your application here
-		}
-	}
-}
-
